lldp: no longer keep parsed attributes in LldpNeighbor
We only need to parse them to construct the GVariant. There is
no need to keep them around otherwise.

We still keep LldpAttrs array and don't construct the GVariant right
away. The benefit is that this way while parsing we set the array
fields, and afterwards, when we generate the string dictionary, the
keys are sorted.