lldp: delay change notification for LLDP neighbor events
We already rate limit change events by two seconds. When we notice
that something changed, we call data_changed_schedule().

Previously, that would immediately issue the change notification,
if ratelimiting currently was not in effect. That means, if we happen
go receive two LLDP neighbor events in short succession, then the
first one will trigger the change notification right away, while
the second will be rate limited.

Avoid that by always issue scheduling the change notification in
the background. And if we currently are not rate limited, with
an idle handler with low priority.